fix: resolve all critical/high technical debt findings
Stub/fallback routes (9 files):
- Replace empty dataset fallbacks with 503 Service Unavailable responses
- Add serviceUnavailableResponse() helper in lib/api/server.ts
- Update 7 test files to expect 503 instead of 200+empty

Large file splits:
- types.ts (1115 lines) → 8 domain files + barrel re-export
- gameday-repository.ts (1563 lines) → 4 sub-repositories + facade
- participant.ts (1611 lines) → 5 sub-router files by domain

Test quality:
- Split auth test assertion roulette into 4 focused tests
- Extract shared callSessionCallback() helper

Security:
- Add NODE_ENV !== 'production' guard to leaderboard-service AUTH_SKIP
