+++ apps/application-plane/app/api/admin/participants/__tests__/route.test.ts
@@ -19,6 +19,10 @@ vi.mock('@/lib/api/server', () => ({
     new Response(JSON.stringify({ error: msg }), { status: 403 }),
   badRequestResponse: (msg = 'Bad Request') =>
     new Response(JSON.stringify({ error: msg }), { status: 400 }),
+  serviceUnavailableResponse: (msg = 'Service unavailable') =>
+    new Response(JSON.stringify({ error: msg, statusCode: 503 }), {
+      status: 503,
+    }),
   successResponse: <T>(data: T, status = 200) =>
     new Response(JSON.stringify(data), { status }),
 }));
@@ -62,7 +66,7 @@ describe('Admin Participants API', () => {
       await expect(response.json()).resolves.toEqual(mockParticipants);
     });
 
-    it('AUTH_SKIP の Unauthorized は空一覧にフォールバックすべき', async () => {
+    it('AUTH_SKIP の Unauthorized は 503 を返すべき', async () => {
       const session: Session = {
         user: { name: 'Admin', email: 'admin@example.com' },
         expires: new Date().toISOString(),
@@ -77,16 +81,13 @@ describe('Admin Participants API', () => {
       );
       const response = await GET(request);
 
-      expect(response.status).toBe(200);
-      await expect(response.json()).resolves.toEqual({
-        participants: [],
-        total: 0,
-        page: 2,
-        pageSize: 25,
+      expect(response.status).toBe(503);
+      await expect(response.json()).resolves.toMatchObject({
+        error: 'Failed to fetch participants',
       });
     });
 
-    it('network error は空一覧にフォールバックすべき', async () => {
+    it('network error は 503 を返すべき', async () => {
       const session: Session = {
         user: { name: 'Admin', email: 'admin@example.com' },
         expires: new Date().toISOString(),
@@ -101,12 +102,9 @@ describe('Admin Participants API', () => {
       );
       const response = await GET(request);
 
-      expect(response.status).toBe(200);
-      await expect(response.json()).resolves.toEqual({
-        participants: [],
-        total: 0,
-        page: 1,
-        pageSize: 10,
+      expect(response.status).toBe(503);
+      await expect(response.json()).resolves.toMatchObject({
+        error: 'Failed to fetch participants',
       });
     });
   });
